Update lvgl library version and implement GUI service
The lvgl library has been downgraded from v9.1.0 to v8.4.0. The GUI service is introduced and its basic setup methods like begin, setup, and loop are implemented in GuiService.h and GuiService.cpp. The setup of the display and the drawing buffer for the GUI is also added. Some validation on receiving and transmitting radio messages are also introduced in RadioService.cpp. In ServiceBase.h, the statement in softAssert function has been corrected.