Add keyboard and touch functionality to GuiService
This commit adds methods to handle keyboard and touch events to GuiService. It includes the registration of touch and keyboard devices to GuiService, allocation of new touch and keyboard drivers, and implementations of handleTouch and handleKeyBoard functions. Along with these, changes have been made to KeyboardService to return lvgl compatible key codes and lastly, callback registration for touch and keyboard handlers has been added to main setup function.